pyproject.toml
[project]
name = "keychain_login"
version = "0.1.0"
description = "Very Crude Flask Keychain Login"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "hive-nectar>=0.0.7",
]
